refact: dry tags field setter
Also removes HTML sanitizer from tags.
