+++ server/models/Expense.js
@@ -5,8 +5,7 @@ import status from '../constants/expense_status';
 import expenseType from '../constants/expense_type';
 import CustomDataTypes from '../models/DataTypes';
 import { reduceArrayToCurrency } from '../lib/currency';
-import { stripHTML } from '../lib/sanitize-html';
-import { validateTags } from '../lib/tags';
+import { validateTags, setTags } from '../lib/tags';
 import models, { Op } from './';
 import { PayoutMethodTypes } from './PayoutMethod';
 
@@ -144,27 +143,7 @@ export default function (Sequelize, DataTypes) {
 
       tags: {
         type: DataTypes.ARRAY(DataTypes.STRING),
-        set(tags) {
-          if (tags) {
-            tags = tags
-              .map(tag => {
-                if (tag) {
-                  const upperCase = tag.toUpperCase();
-                  const cleanTag = upperCase.trim().replace(/\s+/g, ' ');
-                  return stripHTML(cleanTag);
-                }
-              })
-              .filter(tag => {
-                return tag && tag.length > 0;
-              });
-          }
-
-          if (!tags || tags.length === 0) {
-            this.setDataValue('tags', null);
-          } else if (tags) {
-            this.setDataValue('tags', Array.from(new Set(tags)));
-          }
-        },
+        set: setTags,
         validate: { validateTags },
       },
     },
